Refactor naming and  directory structure
