+++ cmd/registry/cmd/compute/conformance/lint_factory.go
@@ -1,4 +1,4 @@
-package core
+package conformance
 
 import (
 	"errors"
@@ -7,7 +7,9 @@ import (
 
 func CreateLinter(linter_name string) (Linter, error) {
 	if linter_name == "spectral" {
-		return SpectralLinter{Name: "spectral", Rules: make(map[string][]string)}, nil
+		return SpectralLinter{Rules: make(map[string][]string)}, nil
+	} else if linter_name == "api-linter" {
+		return SpectralLinter{Rules: make(map[string][]string)}, nil
 	}
 
 	return nil, errors.New(fmt.Sprintf("Unknown Linter: %s", linter_name))
