+++ cmd/registry/cmd/compute/conformance/linter_plugin_iface.go
@@ -1,10 +1,9 @@
-package core
+package conformance
 
 import "github.com/apigee/registry/rpc"
 
 type Linter interface {
 	AddRule(mimeType string, rule string)
-	GetRules(mimeType string) []string
 	GetName() string
 	SupportsMimeType(mimeType string) bool
 	Lint(mimeType string, path string) (*rpc.Lint, error)
